Refactor form and auth module names for consistency
