+++ web/src/pages/auth/mod.rs
@@ -1,13 +1,13 @@
 pub mod confirm_email;
 pub mod forgot_password;
-pub mod login;
+pub mod log_in;
 pub mod set_password;
-pub mod signup;
+pub mod sign_up;
 pub mod verify_forgot_password;
 
 pub use confirm_email::*;
 pub use forgot_password::*;
-pub use login::*;
+pub use log_in::*;
 pub use set_password::*;
-pub use signup::*;
+pub use sign_up::*;
 pub use verify_forgot_password::*;
