+++ web/styles/pages/auth/_sign-up.scss
@@ -1,0 +1,2 @@
+.sign-up-page {
+}
