+++ web/styles/pages/auth/index.scss
@@ -1,0 +1,1 @@
+@forward "sign-up";
