+++ web/styles/pages/index.scss
@@ -1,1 +1,2 @@
 @forward "home";
+@forward "auth";
